requirements.txt.txt
streamlit
pandas
opencv-python-headless
pytesseract
scikit-learn
pillow
